Added addition tests with complex selenium options on `Open Browser` keyword
Put in and triaged a couple of tests that came from a question wiuthin the slack channels. Can one convert a complex selenium
options argument that uses a dictionary into a similarly complex options which just uses the options string and get the same resulting
browser? Triaging at th moemnt as these need more asertations to verify they are the same.