Added press keys test to demonstrate two keystroke delete
A question was raised in the repo issue noting an apparent failure to clear out a field. The user used
`COMMAND+A+BACKSPACE` to do this. This is incorrect in that this does a single combined kepypress, pressing
the backspace while holding/pressing the command+a. Instead it should be two separate keystrokes. First
a the select all (`COMMAND+A`) and then in a separate argument the delete, as in

```
Press Keys    COMMAND+A    BACKSPACE
```

This commit always shows an interesting issue with getting the changed text from the text area. In this
commit one should see the commented out Element Text Should Be calls. Their removal is due to the fact
these were failing. The text was remaining the same. There are a couple reasons why this was the case.
First for text area I believe the text is in the value property instead of the text. Second is that
one must query the property as a property can change whereas the DOM attribute is that which is coded.
So if one replaces `Get Property` with `Get DOM Attribute` the test will fail again as the DOM attribute
does not change. it is what the page was coded with. But the value property does.

This is an interesting exercise in learning a few aspects of html, selenium, and debugging.